fixen die aufgaben, 5-1 und 5-2 von 27-10-2025

+++ 2025-10-27/aufgabe5-1.c
@@ -13,13 +13,13 @@ int main(void){
 
     // Deklaration der Variablen
     
-    int inputZahl;
+    float inputZahl;
 
 
     // Eingabeaufforderung 
 
     printf("Bitte eine Zahl eingeben: ");
-    scanf("%d", &inputZahl);
+    scanf("%f", &inputZahl);
 
 
 // Überprüfung, ob die Zahl negativ ist
@@ -27,7 +27,7 @@ int main(void){
         printf("Fehler: Die eingegebene Zahl ist negativ. Die Wurzel einer negativen Zahl ist nicht definiert.\n");
     } else {
 
-        printf("Die Wurzel aus %d ist %.2f\n", inputZahl, sqrt(inputZahl));
+        printf("Die Wurzel aus %f ist %f\n", inputZahl, sqrt(inputZahl));
 
     }
